requirements.txt
streamlit
torch
torchvision
Pillow
streamlit-drawable-canvas
psycopg2
